+++ sdk/iot_hub/src/service/requests/mod.rs
@@ -8,7 +8,7 @@ mod get_configuration;
 mod get_identity;
 mod get_twin;
 mod invoke_method;
-mod query_builder;
+mod query;
 mod update_or_replace_twin_builder;
 
 pub use apply_on_edge_device::ApplyOnEdgeDeviceBuilder;
@@ -21,5 +21,5 @@ pub(crate) use get_configuration::GetConfigurationBuilder;
 pub(crate) use get_identity::get_identity;
 pub(crate) use get_twin::get_twin;
 pub use invoke_method::InvokeMethodBuilder;
-pub use query_builder::QueryBuilder;
+pub use query::QueryBuilder;
 pub use update_or_replace_twin_builder::UpdateOrReplaceTwinBuilder;
